policy

+++ app/Http/Controllers/CustomersController.php
@@ -32,6 +32,8 @@ class CustomersController extends Controller
 
     public function store()
     {
+        $this->authorize('create', Customer::class);
+
         $customer = Customer::create($this->validateRequest());
 
         $this->storeImage($customer);
@@ -63,6 +65,7 @@ class CustomersController extends Controller
 
     public function destroy(Customer $customer)
     {
+        $this->authorize('create', $customer);
         $customer->delete();
 
         return redirect('customers');
